requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
aiosqlite>=0.17.0
asyncpg>=0.28.0
Pillow>=10.0.0
duckduckgo-search>=4.0.0
numexpr>=2.8.0
httpx>=0.24.0